Remove bug report button from navbar
