+++ app/web/service/bugs.ts
@@ -1,4 +1,4 @@
-import { BugReportFormData } from "components/Navigation/ReportButton";
+import { BugReportFormData } from "components/Navigation/ReportDialog";
 import { ReportBugReq, ScreenResolution } from "proto/bugs_pb";
 
 import client from "./client";
